requirements.txt
Pillow>=11.1.0
pytesseract==0.3.13
